refactor: index Page  (#50)
* fix: 파일 이름 오타 수정

* fix: javascript naming 규칙에 의해 변수 이름 수정

* fix: Menu type 이름 수정

* refactor: sidebar 이름 겹침에 따른 수정 및 타입 추가

* fix: menu props -> sidebarmenu props
