+++ components/sidebar/index.tsx
@@ -1,8 +1,8 @@
-import { Container } from '@/components/sidebar/styled';
-import { Menu } from '@/components/sidebar/menu';
-import { MenuProps } from '@/shared/types/menu';
+import SideBarMenu from './menu';
+import { Container } from './styles';
+import { Menu } from '@/shared/types/menu';
 
-const MenuList: MenuProps[] = [
+const menuList = [
   {
     title: 'Inbox',
     iconUrl: 'https://user-images.githubusercontent.com/51700274/178726362-ece91e50-ab92-48ca-aec0-54b92c021997.png',
@@ -18,9 +18,9 @@ const MenuList: MenuProps[] = [
 const SideBar = () => {
   return (
     <Container>
-      {MenuList.map((menu: MenuProps, index: number) => (
+      {menuList.map((menu: Menu, index: number) => (
         // !TODO index는 키이면 안되므로 다음에 수정 필요
-        <Menu key={index} title={menu.title} iconUrl={menu.iconUrl} link={menu.link} />
+        <SideBarMenu key={index} title={menu.title} iconUrl={menu.iconUrl} link={menu.link} />
       ))}
     </Container>
   );
