+++ shared/types/menu.ts
@@ -1,4 +1,4 @@
-export interface MenuProps {
+export interface Menu {
   title: string;
   iconUrl: string;
   link: string;
